backend/requirements.txt
# Python Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
pyodbc==5.2.0

# Azure Services
azure-storage-blob==12.19.0
azure-cognitiveservices-speech==1.33.0
openai==1.3.5

# PDF Generation
reportlab==4.0.7

# Data Processing
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
annotated-types==0.7.0